# Conflicts:
#	service/src/main/java/com/epam/rft/atsy/service/impl/CandidateServiceImpl.java
#	web/src/main/resources/i18n/messages_en.properties
#	web/src/main/resources/i18n/messages_hu.properties
#	web/src/main/webapp/WEB-INF/pages/candidates.jsp

+++ service/src/main/java/com/epam/rft/atsy/service/domain/CandidateDTO.java
@@ -23,14 +23,14 @@ public class CandidateDTO {
 
     private String referer;
 
-    private Byte languageSkill;
+    private Short languageSkill;
 
     private String description;
 
     public CandidateDTO() {
     }
 
-    public CandidateDTO(String name, String email, String phone, String description, String referer, Byte languageSkill) {
+    public CandidateDTO(String name, String email, String phone, String description, String referer, Short languageSkill) {
         this.name = name;
         this.email = email;
         this.phone = phone;
@@ -79,11 +79,11 @@ public class CandidateDTO {
         this.referer = referer;
     }
 
-    public int getLanguageSkill() {
+    public Short getLanguageSkill() {
         return languageSkill;
     }
 
-    public void setLanguageSkill(Byte languageSkill) {
+    public void setLanguageSkill(Short languageSkill) {
         this.languageSkill = languageSkill;
     }
